refactor QrCodeScannerDialog into its own widget, use thru Frost process
